refactor(mdx): create createFusumaProps
