+++ packages/mdx-loader/src/mdxPlugin.js
@@ -5,56 +5,8 @@ const visit = require('unist-util-visit');
 
 const mdxAstToMdxHast = require('@mdx-js/mdx/mdx-ast-to-mdx-hast');
 const { toJSX } = require('@mdx-js/mdx/mdx-hast-to-jsx');
+const createFusumaProps = require('./createFusumaProps');
 const transferMarkdownImageNodeToJSX = require('./transferMarkdownImageNodeToJSX');
-
-function createFusumaProps(nodes) {
-  const property = {};
-
-  nodes.forEach(({ type, value }) => {
-    if (type === 'comment') {
-      const v = value.trim();
-
-      if (v === 'contents') {
-        property.contents = true;
-      }
-      if (v === 'screen') {
-        property.screen = true;
-      }
-      if (v.slice(0, 4) === 'note') {
-        // sanitize
-        const escapeMap = {
-          '&': '&amp;',
-          '<': '&lt;',
-          '>': '&gt;',
-          '"': '&quot;',
-          "'": '&#x27;',
-          // '/': '&#x2F;'
-        };
-
-        const s = v
-          .slice(4)
-          .trim()
-          .replace(/[&<>"']/gim, (m) => escapeMap[m])
-          .replace(/\n/g, '\\n');
-
-        property.note = s;
-      }
-      if (v.slice(0, 8) === 'classes:') {
-        property.classes = v.slice(8).trim();
-      }
-      if (v.slice(0, 13) === 'sectionTitle:') {
-        property.sectionTitle = v.slice(13).trim();
-      }
-      // if (v.slice(0, 5) === 'line:') {
-      //   property.lines = v.slice(13).trim();
-      // }
-    }
-  });
-
-  return `{${Object.entries(property)
-    .map(([key, value]) => `${key}: '${value}'`)
-    .join(',')}}`;
-}
 
 function mdxPlugin() {
   return (tree) => {
